% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdf.R
\name{ensure_type}
\alias{ensure_type}
\title{S3 generic to ensure we don't change data type when assigning into column}
\usage{
ensure_type(x, value)
}
\arguments{
\item{x}{column values}

\item{value}{new value}
}
\value{
coerced new value
}
\description{
S3 generic to ensure we don't change data type when assigning into column
}

