\docType{data}
\name{GComponentObservable}
\alias{GComponentObservable}
\title{GComponentObservable adds the observable interface}
\format{Reference class 'refObjectGenerator' [package "methods"] with 2 fields
 and 17 methods, of which 6 are possibly relevant:
   accessors, fields, help, lock, methods, new}
\usage{
  GComponentObservable
}
\description{
  GComponentObservable adds the observable interface
}
\keyword{dataset}

