\docType{data}
\name{GEdit}
\alias{GEdit}
\title{The GEdit class adds some methods beyond the spec: \code{set_error}, \code{clear_error}, \code{validate_value}}
\format{Reference class 'refObjectGenerator' [package "methods"] with 2 fields
 and 17 methods, of which 6 are possibly relevant:
   accessors, fields, help, lock, methods, new}
\usage{
  GEdit
}
\description{
  The GEdit class adds some methods beyond the spec:
  \code{set_error}, \code{clear_error},
  \code{validate_value}
}
\keyword{dataset}

