\docType{data}
\name{GMessage}
\alias{GMessage}
\title{subclass for message dialog}
\format{Reference class 'refObjectGenerator' [package "methods"] with 2 fields
 and 17 methods, of which 6 are possibly relevant:
   accessors, fields, help, lock, methods, new}
\usage{
  GMessage
}
\description{
  subclass for message dialog
}
\keyword{dataset}

