\name{GTree}
\alias{GTree}
\title{Class for gtree objects. Extra reference methods are \code{set_multiple} to set whether multiple or single selection is being used.}
\description{
  Class for gtree objects. Extra reference methods are
  \code{set_multiple} to set whether multiple or single
  selection is being used.
}

