\name{.addStockIcons.guiWidgetsToolkitRGtk2}
\alias{.addStockIcons.guiWidgetsToolkitRGtk2}
\alias{.getStockIconByName.guiWidgetsToolkitRGtk2}
\alias{.getStockIcons.guiWidgetsToolkitRGtk2}
\alias{addToGtkStockIcons}
\title{add stock icons}
\usage{
  .addStockIcons.guiWidgetsToolkitRGtk2(toolkit, iconNames,
  iconFiles, ...)

  .getStockIcons.guiWidgetsToolkitRGtk2(toolkit, ...)

  .getStockIconByName.guiWidgetsToolkitRGtk2(toolkit, name,
  ...)

  addToGtkStockIcons(iconNames, iconFiles)
}
\description{
  add stock icons

  Returns list of stock ids

  return stock id

  helper function
}

