\name{getWidget.RGtkObject}
\alias{.ggroup.guiWidgetsToolkitRGtk2}
\alias{.gwindow.guiWidgetsToolkitRGtk2}
\alias{getBlock.RGtkObject}
\alias{getWidget.RGtkObject}
\title{method for stopping getWidget}
\usage{
  getWidget.RGtkObject(obj)

  getBlock.RGtkObject(obj)

  .gwindow.guiWidgetsToolkitRGtk2(toolkit, title, visible =
  visible, name, width, height, parent, handler, action,
  ...)

  .ggroup.guiWidgetsToolkitRGtk2(toolkit, horizontal =
  TRUE, spacing = 5, use.scrollwindow = FALSE, container =
  NULL, ...)
}
\description{
  method for stopping getWidget

  S3 method for stopping getBlock

  toolkit constructor for gwindow

  toolkit constructor for ggroup
}

