% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtk-misc.R
\name{isRightMouseClick}
\alias{isRightMouseClick}
\title{Return TRUE/FALSE if right mouse click}
\usage{
isRightMouseClick(e)
}
\arguments{
\item{e}{event for mouse press}
}
\value{
TRUE or FALSE
}
\description{
To be called from key-press|release-event
}

