\name{GEdit}
\alias{GEdit}
\title{The GEdit class adds some methods beyond the spec: \code{set_error}, \code{clear_error}, \code{validate_value}}
\usage{
  GEdit(...)
}
\arguments{
  \item{...}{passed to constructor}
}
\description{
  The GEdit class adds some methods beyond the spec:
  \code{set_error}, \code{clear_error},
  \code{validate_value}
}

