\docType{data}
\name{GTree}
\alias{GTree}
\title{Base class}
\format{Reference class 'refObjectGenerator' [package "methods"] with 2 fields
 and 17 methods, of which 6 are possibly relevant:
   accessors, fields, help, lock, methods, new}
\usage{
  GTree
}
\description{
  For \code{GTree}, there are extra reference methods:
  \code{set_multiple} to set whether multiple or single
  selection is being used.
}
\keyword{dataset}

