\name{gWidgets2RGtk2-package}
\alias{gWidgets2RGtk2-package}
\alias{gWidgets2RGtk2}
\docType{package}
\title{
Toolkit implementation of gWidgets API
}
\description{
The gWidgets2 API provides an abstract means to interact with an underlying graphical toolkit library. This package provides the connectino between gWidgets2  and RGtk2.
}
\author{
John Verzani <jverzani@gmail.com>
}
\references{
http://CRAN.R-project.org/doc/Rnews/Rnews_2007-3.pdf
}
\keyword{ package }
