% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtk-misc.R, R/aaa.R, R/dialogs.R,
%   R/gaction.R, R/gbutton.R, R/gcalendar.R, R/gcheckbox.R, R/gcheckboxgroup.R,
%   R/gcombobox.R, R/gtable.R, R/gmenu.R, R/gdf.R, R/gedit.R, R/ggroup.R,
%   R/gframe.R, R/gexpandgroup.R, R/gfile.R, R/gformlayout.R, R/ggraphics.R,
%   R/gimage.R, R/glabel.R, R/glayout.R, R/gnotebook.R, R/gpanedgroup.R,
%   R/gprogressbar.R, R/gradio.R, R/gseparator.R, R/gslider.R, R/gspinbutton.R,
%   R/gstackwidget.R, R/gstatusbar.R, R/gtext.R, R/gtimer.R, R/gtoolbar.R,
%   R/gtree.R, R/gvarbrowser.R, R/gwindow.R, R/icons.R
\name{getWidget.RGtkObject}
\alias{getWidget.RGtkObject}
\alias{getBlock.RGtkObject}
\alias{.docstub}
\alias{.gmessage.guiWidgetsToolkitRGtk2}
\alias{.gconfirm.guiWidgetsToolkitRGtk2}
\alias{.ginput.guiWidgetsToolkitRGtk2}
\alias{.gbasicdialog.guiWidgetsToolkitRGtk2}
\alias{.galert.guiWidgetsToolkitRGtk2}
\alias{.gaction.guiWidgetsToolkitRGtk2}
\alias{.gbutton.guiWidgetsToolkitRGtk2}
\alias{.gcalendar.guiWidgetsToolkitRGtk2}
\alias{.gcheckbox.guiWidgetsToolkitRGtk2}
\alias{.gcheckboxgroup.guiWidgetsToolkitRGtk2}
\alias{.gcombobox.guiWidgetsToolkitRGtk2}
\alias{.gtable.guiWidgetsToolkitRGtk2}
\alias{.gmenu.guiWidgetsToolkitRGtk2}
\alias{.gdf.guiWidgetsToolkitRGtk2}
\alias{.gedit.guiWidgetsToolkitRGtk2}
\alias{.ggroup.guiWidgetsToolkitRGtk2}
\alias{.gframe.guiWidgetsToolkitRGtk2}
\alias{.gexpandgroup.guiWidgetsToolkitRGtk2}
\alias{.gfile.guiWidgetsToolkitRGtk2}
\alias{.gfilebrowse.guiWidgetsToolkitRGtk2}
\alias{.gformlayout.guiWidgetsToolkitRGtk2}
\alias{.ggraphics.guiWidgetsToolkitRGtk2}
\alias{.gimage.guiWidgetsToolkitRGtk2}
\alias{.glabel.guiWidgetsToolkitRGtk2}
\alias{.glayout.guiWidgetsToolkitRGtk2}
\alias{.gnotebook.guiWidgetsToolkitRGtk2}
\alias{.gpanedgroup.guiWidgetsToolkitRGtk2}
\alias{.gprogressbar.guiWidgetsToolkitRGtk2}
\alias{.gradio.guiWidgetsToolkitRGtk2}
\alias{.gseparator.guiWidgetsToolkitRGtk2}
\alias{.gslider.guiWidgetsToolkitRGtk2}
\alias{.gspinbutton.guiWidgetsToolkitRGtk2}
\alias{.gstackwidget.guiWidgetsToolkitRGtk2}
\alias{.gstatusbar.guiWidgetsToolkitRGtk2}
\alias{.gtext.guiWidgetsToolkitRGtk2}
\alias{.gtimer.guiWidgetsToolkitRGtk2}
\alias{.gtoolbar.guiWidgetsToolkitRGtk2}
\alias{.gtree.guiWidgetsToolkitRGtk2}
\alias{.gvarbrowser.guiWidgetsToolkitRGtk2}
\alias{.gwindow.guiWidgetsToolkitRGtk2}
\alias{.addStockIcons.guiWidgetsToolkitRGtk2}
\alias{.getStockIcons.guiWidgetsToolkitRGtk2}
\alias{.getStockIconByName.guiWidgetsToolkitRGtk2}
\alias{addToGtkStockIcons}
\alias{.stockIconFromObject.guiWidgetsToolkitRGtk2}
\title{method for stopping getWidget}
\usage{
\method{getWidget}{RGtkObject}(obj)

\method{getBlock}{RGtkObject}(obj)

.docstub()

\method{.gmessage}{guiWidgetsToolkitRGtk2}(toolkit, msg, title = "message",
  icon = c("info", "warning", "error", "question"), parent = NULL, ...)

\method{.gconfirm}{guiWidgetsToolkitRGtk2}(toolkit, msg, title = "Confirm",
  icon = c("info", "warning", "error", "question"), parent = NULL, ...)

\method{.ginput}{guiWidgetsToolkitRGtk2}(toolkit, msg, text = "",
  title = "Input", icon = c("info", "warning", "error", "question"),
  parent = NULL, ...)

\method{.gbasicdialog}{guiWidgetsToolkitRGtk2}(toolkit, title = "Dialog",
  parent = NULL, do.buttons = TRUE, handler = NULL, action = NULL, ...)

\method{.galert}{guiWidgetsToolkitRGtk2}(toolkit, msg, title = "message",
  delay = 3, parent = NULL, ...)

\method{.gaction}{guiWidgetsToolkitRGtk2}(toolkit, label, tooltip = NULL,
  icon = NULL, key.accel = NULL, handler = NULL, action = NULL,
  parent = NULL, ...)

\method{.gbutton}{guiWidgetsToolkitRGtk2}(toolkit, text, handler, action,
  container, ...)

\method{.gcalendar}{guiWidgetsToolkitRGtk2}(toolkit, text = "",
  format = "\%Y-\%m-\%d", handler = NULL, action = NULL,
  container = NULL, ...)

\method{.gcheckbox}{guiWidgetsToolkitRGtk2}(toolkit, text, checked = FALSE,
  use.togglebutton = FALSE, handler = NULL, action = NULL,
  container = NULL, ...)

\method{.gcheckboxgroup}{guiWidgetsToolkitRGtk2}(toolkit = NULL, items,
  checked = FALSE, horizontal = FALSE, use.table = FALSE,
  handler = NULL, action = NULL, container = NULL, ...)

\method{.gcombobox}{guiWidgetsToolkitRGtk2}(toolkit, items, selected = 1,
  editable = FALSE, coerce.with = NULL, handler = NULL, action = NULL,
  container = NULL, ...)

\method{.gtable}{guiWidgetsToolkitRGtk2}(toolkit, items, multiple = FALSE,
  chosen.col = 1, icon.col = NULL, tooltip.col = NULL, handler = NULL,
  action = NULL, container = NULL, ...)

\method{.gmenu}{guiWidgetsToolkitRGtk2}(toolkit, menu.list = list(),
  popup = FALSE, container = NULL, ...)

\method{.gdf}{guiWidgetsToolkitRGtk2}(toolkit, items = NULL, handler = NULL,
  action = NULL, container = NULL, ...)

\method{.gedit}{guiWidgetsToolkitRGtk2}(toolkit, text = "", width = 25,
  coerce.with = NULL, initial.msg = initial.msg, handler = NULL,
  action = NULL, container = NULL, ...)

\method{.ggroup}{guiWidgetsToolkitRGtk2}(toolkit, horizontal = TRUE,
  spacing = 5, use.scrollwindow = FALSE, container = NULL, ...)

\method{.gframe}{guiWidgetsToolkitRGtk2}(toolkit, text, markup, pos,
  horizontal = TRUE, spacing = 5, container = NULL, ...)

\method{.gexpandgroup}{guiWidgetsToolkitRGtk2}(toolkit, text, markup,
  horizontal = TRUE, handler = NULL, action = NULL, container = NULL,
  ...)

\method{.gfile}{guiWidgetsToolkitRGtk2}(toolkit, text = "", type = c("open",
  "save", "selectdir"), initial.filename = NULL, initial.dir = getwd(),
  filter = list(), multi = FALSE, ...)

\method{.gfilebrowse}{guiWidgetsToolkitRGtk2}(toolkit, text = "",
  type = c("open", "save", "selectdir"), initial.filename = NULL,
  initial.dir = getwd(), filter = list(), quote = TRUE, handler = NULL,
  action = NULL, container = NULL, ...)

\method{.gformlayout}{guiWidgetsToolkitRGtk2}(toolkit, align = "left",
  spacing = 5, container = NULL, ...)

\method{.ggraphics}{guiWidgetsToolkitRGtk2}(toolkit, width = dpi * 6,
  height = dpi * 6, dpi = 75, ps = 12, handler = NULL, action = NULL,
  container = NULL, ...)

\method{.gimage}{guiWidgetsToolkitRGtk2}(toolkit, filename = "",
  dirname = "", stock.id = NULL, size = "", handler = NULL,
  action = NULL, container = NULL, ...)

\method{.glabel}{guiWidgetsToolkitRGtk2}(toolkit, text = "", markup = FALSE,
  editable = FALSE, handler = NULL, action = NULL, container = NULL,
  ...)

\method{.glayout}{guiWidgetsToolkitRGtk2}(toolkit, homogeneous = FALSE,
  spacing = 10, container = NULL, ...)

\method{.gnotebook}{guiWidgetsToolkitRGtk2}(toolkit, tab.pos = 3,
  container = NULL, ...)

\method{.gpanedgroup}{guiWidgetsToolkitRGtk2}(toolkit, horizontal = TRUE,
  container = NULL, ...)

\method{.gprogressbar}{guiWidgetsToolkitRGtk2}(toolkit, value, container, ...)

\method{.gradio}{guiWidgetsToolkitRGtk2}(toolkit, items, selected = 1,
  horizontal = FALSE, handler = NULL, action = NULL, container = NULL,
  ...)

\method{.gseparator}{guiWidgetsToolkitRGtk2}(toolkit, horizontal = TRUE,
  container = NULL, ...)

\method{.gslider}{guiWidgetsToolkitRGtk2}(toolkit, from = 0, to = 100,
  by = 1, value = from, horizontal = TRUE, handler = NULL,
  action = NULL, container = NULL, ...)

\method{.gspinbutton}{guiWidgetsToolkitRGtk2}(toolkit, from = 0, to = 10,
  by = 1, value = from, digits = 0, handler = NULL, action = NULL,
  container = NULL, ...)

\method{.gstackwidget}{guiWidgetsToolkitRGtk2}(toolkit, container = NULL, ...)

\method{.gstatusbar}{guiWidgetsToolkitRGtk2}(toolkit, text = "",
  container = NULL, ...)

\method{.gtext}{guiWidgetsToolkitRGtk2}(toolkit, text = NULL, width = NULL,
  height = 300, font.attr = NULL, wrap = TRUE, handler = NULL,
  action = NULL, container = NULL, ...)

\method{.gtimer}{guiWidgetsToolkitRGtk2}(toolkit, ms, FUN, data = NULL,
  one.shot = FALSE, start = TRUE)

\method{.gtoolbar}{guiWidgetsToolkitRGtk2}(toolkit, toolbar.list = list(),
  style = c("both", "icons", "text", "both-horiz"), container = NULL, ...)

\method{.gtree}{guiWidgetsToolkitRGtk2}(toolkit, offspring = NULL,
  offspring.data = NULL, chosen.col = 1, offspring.col = 2,
  icon.col = NULL, tooltip.col = NULL, multiple = FALSE, handler = NULL,
  action = NULL, container = NULL, ...)

\method{.gvarbrowser}{guiWidgetsToolkitRGtk2}(toolkit, handler = NULL,
  action = "summary", container = NULL, ...)

\method{.gwindow}{guiWidgetsToolkitRGtk2}(toolkit, title, visible = visible,
  name, width, height, parent, handler, action, ...)

\method{.addStockIcons}{guiWidgetsToolkitRGtk2}(toolkit, iconNames, iconFiles,
  ...)

\method{.getStockIcons}{guiWidgetsToolkitRGtk2}(toolkit, ...)

\method{.getStockIconByName}{guiWidgetsToolkitRGtk2}(toolkit,name, ...)

addToGtkStockIcons(iconNames, iconFiles)

\method{.stockIconFromObject}{guiWidgetsToolkitRGtk2}(toolkit,obj, ...)
}
\arguments{
\item{obj}{R object to get icon from}

\item{toolkit}{undocumented argument, cf. gWidgets2 documentation.}

\item{msg}{undocumented argument, cf. gWidgets2 documentation.}

\item{title}{undocumented argument, cf. gWidgets2 documentation.}

\item{icon}{undocumented argument, cf. gWidgets2 documentation.}

\item{parent}{undocumented argument, cf. gWidgets2 documentation.}

\item{...}{undocumented argument, cf. gWidgets2 documentation.}

\item{text}{undocumented argument, cf. gWidgets2 documentation.}

\item{do.buttons}{undocumented argument, cf. gWidgets2 documentation.}

\item{handler}{undocumented argument, cf. gWidgets2 documentation.}

\item{action}{undocumented argument, cf. gWidgets2 documentation.}

\item{delay}{undocumented argument, cf. gWidgets2 documentation.}

\item{label}{undocumented argument, cf. gWidgets2 documentation.}

\item{tooltip}{undocumented argument, cf. gWidgets2 documentation.}

\item{key.accel}{undocumented argument, cf. gWidgets2 documentation.}

\item{container}{undocumented argument, cf. gWidgets2 documentation.}

\item{format}{undocumented argument, cf. gWidgets2 documentation.}

\item{checked}{undocumented argument, cf. gWidgets2 documentation.}

\item{use.togglebutton}{undocumented argument, cf. gWidgets2 documentation.}

\item{items}{undocumented argument, cf. gWidgets2 documentation.}

\item{horizontal}{undocumented argument, cf. gWidgets2 documentation.}

\item{use.table}{undocumented argument, cf. gWidgets2 documentation.}

\item{selected}{undocumented argument, cf. gWidgets2 documentation.}

\item{editable}{undocumented argument, cf. gWidgets2 documentation.}

\item{coerce.with}{undocumented argument, cf. gWidgets2 documentation.}

\item{multiple}{undocumented argument, cf. gWidgets2 documentation.}

\item{chosen.col}{undocumented argument, cf. gWidgets2 documentation.}

\item{icon.col}{undocumented argument, cf. gWidgets2 documentation.}

\item{tooltip.col}{undocumented argument, cf. gWidgets2 documentation.}

\item{menu.list}{undocumented argument, cf. gWidgets2 documentation.}

\item{popup}{undocumented argument, cf. gWidgets2 documentation.}

\item{width}{width of device (pixels)}

\item{initial.msg}{undocumented argument, cf. gWidgets2 documentation.}

\item{spacing}{undocumented argument, cf. gWidgets2 documentation.}

\item{use.scrollwindow}{undocumented argument, cf. gWidgets2 documentation.}

\item{markup}{undocumented argument, cf. gWidgets2 documentation.}

\item{pos}{undocumented argument, cf. gWidgets2 documentation.}

\item{type}{undocumented argument, cf. gWidgets2 documentation.}

\item{initial.filename}{undocumented argument, cf. gWidgets2 documentation.}

\item{initial.dir}{undocumented argument, cf. gWidgets2 documentation.}

\item{filter}{undocumented argument, cf. gWidgets2 documentation.}

\item{multi}{undocumented argument, cf. gWidgets2 documentation.}

\item{quote}{undocumented argument, cf. gWidgets2 documentation.}

\item{align}{undocumented argument, cf. gWidgets2 documentation.}

\item{height}{height of device (pixels)}

\item{dpi}{dots per inch}

\item{ps}{pointsize}

\item{filename}{undocumented argument, cf. gWidgets2 documentation.}

\item{dirname}{undocumented argument, cf. gWidgets2 documentation.}

\item{stock.id}{undocumented argument, cf. gWidgets2 documentation.}

\item{size}{undocumented argument, cf. gWidgets2 documentation.}

\item{homogeneous}{undocumented argument, cf. gWidgets2 documentation.}

\item{tab.pos}{undocumented argument, cf. gWidgets2 documentation.}

\item{value}{undocumented argument, cf. gWidgets2 documentation.}

\item{from}{undocumented argument, cf. gWidgets2 documentation.}

\item{to}{undocumented argument, cf. gWidgets2 documentation.}

\item{by}{undocumented argument, cf. gWidgets2 documentation.}

\item{digits}{digits}

\item{font.attr}{undocumented argument, cf. gWidgets2 documentation.}

\item{wrap}{undocumented argument, cf. gWidgets2 documentation.}

\item{ms}{undocumented argument, cf. gWidgets2 documentation.}

\item{FUN}{undocumented argument, cf. gWidgets2 documentation.}

\item{data}{undocumented argument, cf. gWidgets2 documentation.}

\item{one.shot}{undocumented argument, cf. gWidgets2 documentation.}

\item{start}{undocumented argument, cf. gWidgets2 documentation.}

\item{toolbar.list}{undocumented argument, cf. gWidgets2 documentation.}

\item{style}{undocumented argument, cf. gWidgets2 documentation.}

\item{offspring}{undocumented argument, cf. gWidgets2 documentation.}

\item{offspring.data}{undocumented argument, cf. gWidgets2 documentation.}

\item{offspring.col}{undocumented argument, cf. gWidgets2 documentation.}

\item{visible}{undocumented argument, cf. gWidgets2 documentation.}

\item{name}{name of icon}

\item{iconNames}{undocumented argument, cf. gWidgets2 documentation.}

\item{iconFiles}{undocumented argument, cf. gWidgets2 documentation.}
}
\description{
method for stopping getWidget

S3 method for stopping getBlock

The \code{gWidgets} separation of API from toolkit requires the
exporting of "dot" functions from each toolkit package. These are
not documented, as they mirror the definition in the API package,
but must be to get pass R's package check. Hence this stub.

toolkit implementation for gmessage

toolkit implementation for gconfirm

toolkit implmentation of ginput

toolkit implementation

toolkit implementation of galert

Toolkit constructor

toolkit constructor

Toolkit constructor

Toolkit XXX constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit gedit constructor

toolkit constructor for ggroup

gframe constructor

toolkit constructor

Toolkit implementation

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit label constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit  constructor

Toolkit  constructor

Toolkit constructor

Toolkit  constructor

Toolkit XXX constructor

Toolkit constructor

Toolkit constructor

toolkit implementation

S3 method for gtimer

Toolkit constructor

Toolkit constructor

Toolkit constructor

toolkit constructor for gwindow

add stock icons

Returns list of stock ids

return stock id

helper function

return stock id from object
}
\seealso{
The documentation for this is found at \code{\link{gbutton}}.
}
